requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
ast_transform @ git+https://github.com/goodw21985/ApiConductor.git@387e15a51a5358a6be0599f56ee52e1c90b5aa43#subdirectory=AstTransformation
asyncer==0.0.2
beautifulsoup4==4.12.3
bidict==0.23.1
black==24.4.2
blinker==1.8.2
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
chainlit==1.1.202
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
dataclasses-json==0.5.14
Deprecated==1.2.14
faiss-cpu==1.8.0
fastapi==0.110.3
fastapi-socketio==0.0.10
filelock==3.14.0
filetype==1.2.0
Flask==3.0.3
Flask-SocketIO==5.3.6
fonttools==4.52.1
fsspec==2024.5.0
gevent==24.2.1
gevent-websocket==0.10.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
importlib-metadata==7.0.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
Lazify==0.4.0
literalai==0.0.601
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
msal==1.28.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
packaging==23.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
protobuf==4.25.3
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.2
PyTrie==0.4.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==2.7.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.37.2
sympy==1.12
syncer==2.0.3
tbb==2021.12.0
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.1
typing-inspect==0.9.0
typing_extensions==4.11.0
uptrace==1.24.0
urllib3==2.2.1
uvicorn==0.25.0
waitress==3.0.0
watchfiles==0.20.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
zipp==3.18.2
zope.event==5.0
zope.interface==6.4.post1
